Update Shooter Enemy.cs

+++ Enemies/Shooter Enemy.cs
@@ -68,7 +68,7 @@ namespace GwambaPrimeAdventure.Enemy
 				_projectileRotation = _statistics.CircularUse
 					? Quaternion.AngleAxis( Mathf.Atan2( _targetDirection.y, _targetDirection.x ) * Mathf.Rad2Deg - 90F, Vector3.forward )
 					: Quaternion.AngleAxis( _statistics.DirectionAngle * ( _statistics.TurnRay ? transform.localScale.x.CompareTo( 0F ) : 1F ), Vector3.forward );
-			for ( ushort i = 0; _statistics.Projectiles.Length > i; i++ )
+			for ( byte i = 0; _statistics.Projectiles.Length > i; i++ )
 				if ( _statistics.PureInstance )
 					Instantiate(
 						original: _statistics.Projectiles[ i ],
@@ -124,7 +124,7 @@ namespace GwambaPrimeAdventure.Enemy
 					if ( _statistics.TurnRay )
 						_directionCast *= transform.localScale.x.CompareTo( 0F );
 					_castSize = Physics2D.RaycastNonAlloc( _originCast, _directionCast, _detectionRaycasts, _statistics.PerceptionDistance, WorldBuild.CHARACTER_LAYER_MASK );
-					for ( int i = 0; _castSize > i; i++ )
+					for ( byte i = 0; _castSize > i; i++ )
 						if ( _hasTarget = _detectionRaycasts[ i ].collider.TryGetComponent<IDestructible>( out _ ) )
 							break;
 				}
@@ -153,7 +153,7 @@ namespace GwambaPrimeAdventure.Enemy
 				}
 			}
 		}
-		public new bool Hurt( ushort damage )
+		public new bool Hurt( byte damage )
 		{
 			if ( _statistics.ShootDamaged )
 				Animator.SetTrigger( Shoot );
